% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{AnnenBatrute}
\alias{AnnenBatrute}
\title{AnnenBatrute}
\format{
\preformatted{Simple feature collection with 43 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -57722.3 ymin: 6734803 xmax: 1102774 ymax: 7941176
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 43 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 2002-09-15       (730420.3 7782282, 730409.1 7782388, 730355~
 2         64      100000 2009-05-08       (393161 7324670, 392827.4 7325024, 392338 7~
 3         64      100000 2002-09-15       (563521.5 7633773, 564021.6 7634360, 564415~
 4         64      100000 2002-09-15       (499837.6 7606787, 499846.4 7606783, 500613~
 5         64      100000 2013-04-19       (962099.2 7934023, 962724.3 7935269, 963110~
 6         64      100000 2009-06-18       (776682.3 7808699, 776682 7808701, 776590.8~
 7         64      100000 2002-09-15       (232374.7 7067373, 230686.7 7066181, 228555~
 8         64      100000 2009-05-08       (481684.6 7567549, 479999.8 7564808, 478503~
 9         64      100000 2009-05-08       (473630.5 7463267, 472999.8 7463386, 472375~
10         64      100000 2009-11-13       (94358.9 6978876, 94166.9 6978825, 93172.9 ~
# ... with 33 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Samferdsel_GML.gml}
}
\usage{
AnnenBatrute
}
\description{
AnnenBåtrute
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
