% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ApentOmrade}
\alias{ApentOmrade}
\title{ApentOmrade}
\format{
\preformatted{Simple feature collection with 1859 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -76204 ymin: 6450196 xmax: 1113444 ymax: 7939929
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,859 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 2009-05-06       ((282466 6754595, 284987 6753220, 289186.9 6755209, 289205.7 67552~
 2 2009-05-06       ((-31117.89 6706271, -31153.24 6704963, -32425.91 6704575, -32037.~
 3 2011-03-23       ((112509.5 6648336, 113070 6647122, 112269.3 6644879, 111840 66448~
 4 2011-03-23       ((307212.8 6785089, 307198.2 6785176, 307061.1 6785999, 307086.4 6~
 5 NA               ((865903 7688513, 865682.8 7687035, 865565 7686245, 864989 7687058~
 6 NA               ((858210 7685189, 858193.4 7684009, 858585.2 7683062, 858552 76830~
 7 2010-07-21       ((892381.8 7666943, 891906.3 7666732, 891897.7 7666734, 890561.2 7~
 8 NA               ((865365 7853829, 864456 7853049, 863761.3 7853126, 862118 7853309~
 9 NA               ((929699 7853399, 929823 7855467, 930899 7857400, 932423 7856797, ~
10 NA               ((991988 7855459, 991964 7854989, 991827 7854245, 991366.9 7854619~
# ... with 1,849 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
ApentOmrade
}
\description{
ÅpentOmråde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
