% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Arealbrukgrense}
\alias{Arealbrukgrense}
\title{Arealbrukgrense}
\format{
\preformatted{Simple feature collection with 2959 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -59043 ymin: 6455709 xmax: 1113444 ymax: 7913573
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,959 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 2002-09-15       (818772 7824271, 819252 7824992, 821070 782~
 2         64      100000 2002-09-15       (1079046 7808121, 1079099 7808299, 1079384 ~
 3         64      100000 2002-09-15       (1061071 7805289, 1059527 7805362, 1059362 ~
 4         64      100000 2002-09-15       (1080613 7803214, 1080759 7803246, 1081383 ~
 5         64      100000 2002-09-15       (952999 7911499, 951541 7911579, 951142 791~
 6         64      100000 2002-09-15       (919234 7906992, 919533 7907191, 920032 790~
 7         64      100000 2004-07-01       (1e+06 7899531, 999258 7899732, 998687 7898~
 8         64      100000 2002-09-15       (999124 7897323, 999421 7896306, 1e+06 7896~
 9         64      100000 2002-09-15       (1093877 7885671, 1093877 7887752, 1093096 ~
10         64      100000 2002-09-15       (945467 7885350, 946855 7884756, 948350 788~
# ... with 2,949 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Arealbrukgrense
}
\description{
Arealbrukgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
