% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bane}
\alias{Bane}
\title{Bane}
\format{
\preformatted{Simple feature collection with 106 features and 4 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -42446 ymin: 6467440 xmax: 1076891 ymax: 7806691
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 106 x 5
   målemetode nøyaktighet medium oppdateringsdato                              geometry
 *      <int>       <int> <chr>  <date>                                <LINESTRING [m]>
 1         64      100000 U      2011-09-20       (235683 6634536, 235844.6 6634643, 2~
 2         64      100000 T      2012-03-28       (285556.1 6670637, 285902.2 6671698,~
 3         64      100000 U      2012-03-28       (-30553 6732995, -30335.87 6732941, ~
 4         64      100000 T      2012-03-28       (267168.3 6650542, 267311 6650895, 2~
 5         64      100000 T      2012-04-17       (371748 7150284, 371962.4 7149771, 3~
 6         64      100000 T      2012-03-28       (255598 6650050, 255842.8 6649937, 2~
 7         64      100000 T      2012-03-28       (244465.7 6642784, 244817.9 6643195,~
 8         64      100000 T      2012-03-28        (248813.7 6647682, 249009.9 6647662)
 9         64      100000 T      2012-03-28       (256066.1 6649833, 256241.7 6649941,~
10         64      100000 T      2012-03-28       (187931.3 6648911, 188304.3 6648674,~
# ... with 96 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Samferdsel_GML.gml}
}
\usage{
Bane
}
\description{
Bane
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
