% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bilferjestrekning}
\alias{Bilferjestrekning}
\title{Bilferjestrekning}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 145 features and 6 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -74177.12 ymin: 6522919 xmax: 849429.5 ymax: 7904527
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 145 × 7
   målemetode nøyaktighet medium oppdateringsdato vegkategori vegnummer
 *      <int>       <int> <chr>  <date>           <chr>           <int>
 1         64      100000 T      NA               F                 668
 2         64      100000 T      NA               E                  39
 3         64      100000 T      NA               F                  17
 4         64      100000 T      NA               F                 769
 5         64      100000 T      NA               F                 715
 6         64      100000 T      NA               F                 670
 7         64      100000 T      NA               F                  64
 8         64      100000 T      NA               F                  60
 9         64      100000 T      NA               F                 614
10         64      100000 T      NA               F                  NA
# ℹ 135 more rows
# ℹ 1 more variable: senterlinje <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Samferdsel_GML.gml}
}
\usage{
Bilferjestrekning
}
\description{
Bilferjestrekning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
