% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Dataavgrensning}
\alias{Dataavgrensning}
\title{Dataavgrensning}
\format{
\preformatted{Simple feature collection with 621 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 245118.3 ymin: 6521829 xmax: 1114945 ymax: 7828329
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 621 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                              <LINESTRING [m]>
 1 2005-01-01                    (964709.6 7806335, 965465.5 7807009, 966318.4 7807375)
 2 2005-01-01                                        (1113191 7806314, 1112016 7808218)
 3 NA               (1113191 7806314, 1113207 7806288, 1113535 7805757, 1113616 780562~
 4 NA               (950764.2 7805415, 950764 7805415, 950763.5 7805417, 950763.2 7805~
 5 NA                            (964709.6 7806335, 963800.4 7805524, 963670.1 7805408)
 6 2005-01-01                                      (949881.2 7805180, 950751.9 7805414)
 7 NA                                              (949881.2 7805180, 949174.4 7804989)
 8 2005-01-01                    (963035.6 7804924, 963206.4 7804994, 963670.1 7805408)
 9 2005-01-01                      (944000 7804548, 944102.4 7804613, 944164.6 7804652)
10 2005-01-01                                      (962005.1 7804500, 962120.5 7804548)
# ... with 611 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Dataavgrensning
}
\description{
Dataavgrensning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
