% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Elv}
\alias{Elv}
\title{Elv}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 3 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 265025 ymin: 6566414 xmax: 997099 ymax: 7873955
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 3 × 2
  oppdateringsdato                                                                område
* <date>                                                                   <POLYGON [m]>
1 NA               ((287769 6604306, 287136 6604534, 286428 6604382, 285637.9 6604010, …
2 NA               ((994613 7873673, 993516 7873955, 993185 7873440, 992977 7872443, 99…
3 NA               ((269200 6566818, 269044 6567066, 268872 6567452, 268815 6567580, 26…
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Elv
}
\description{
Elv
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
