% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ElvBekk}
\alias{ElvBekk}
\title{ElvBekk}
\format{
\preformatted{Simple feature collection with 2393 features and 4 fields
Geometry type: GEOMETRY
Dimension:     XY
Bounding box:  xmin: -47954 ymin: 6455920 xmax: 1113674 ymax: 7935049
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,393 x 5
   oppdateringsdato vannBredde målemetode nøyaktighet                          geometry
 * <date>                <int>      <int>       <int>                    <GEOMETRY [m]>
 1 NA                        7         NA          NA POLYGON ((287769 6604306, 287718~
 2 2010-07-21                7         NA          NA POLYGON ((994613 7873673, 994441~
 3 NA                        7         NA          NA POLYGON ((269200 6566818, 268594~
 4 2002-09-15                4         64      100000 LINESTRING (468155 7411803, 4679~
 5 2002-09-15                4         64      100000 LINESTRING (516944 7410751, 5177~
 6 2002-09-15                4         64      100000 LINESTRING (474184 7409740, 4748~
 7 2002-09-15                4         64      100000 LINESTRING (529357 7405602, 5284~
 8 2002-09-15                4         64      100000 LINESTRING (489827 7396233, 4891~
 9 2002-09-15                4         64      100000 LINESTRING (516600 7388341, 5167~
10 2002-09-15                4         64      100000 LINESTRING (498640 7376769, 4986~
# ... with 2,383 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
ElvBekk
}
\description{
ElvBekk
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
