% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ElvBekkKant}
\alias{ElvBekkKant}
\title{ElvBekkKant}
\format{
\preformatted{Simple feature collection with 32 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 265025 ymin: 6566414 xmax: 997099 ymax: 7873955
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 32 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 2002-09-15       (992099 7869501, 992497 7870837, 992777 787~
 2         64      100000 2002-09-15       (991721 7858486, 991540 7858660, 991493 785~
 3         64      100000 2002-09-15       (991827 7854245, 991964 7854989, 991988 785~
 4         64      100000 2002-09-15       (991682 7852370, 991680 7852486, 991704 785~
 5         64      100000 2002-09-15       (991864 7850836, 991690 7851885, 991682 785~
 6         64      100000 2002-09-15       (992693 7854538, 992409 7852770, 992446 785~
 7         64      100000 2002-09-15       (991449 7866507, 991951 7867260, 992167 786~
 8         64      100000 2002-09-15       (991988 7855459, 991926 7856386, 991605 785~
 9         64      100000 2002-09-15       (994613 7873673, 994441 7873143, 994343 787~
10         64      100000 2002-09-15       (992797 7850486, 993037 7850008, 993461 784~
# ... with 22 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
ElvBekkKant
}
\description{
ElvBekkKant
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
