% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ElveKant}
\alias{ElveKant}
\title{ElveKant}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 32 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 265025 ymin: 6566414 xmax: 997099 ymax: 7873955
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 32 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 NA               gen             100000 (992099 7869501, 992497 7870837, 992777 7871…
 2 NA               gen             100000 (991721 7858486, 991540 7858660, 991493 7859…
 3 NA               gen             100000 (991827 7854245, 991964 7854989, 991988 7855…
 4 NA               gen             100000 (991682 7852370, 991680 7852486, 991704 7853…
 5 NA               gen             100000 (991864 7850836, 991690 7851885, 991682 7852…
 6 NA               gen             100000 (992693 7854538, 992409 7852770, 992446 7852…
 7 NA               gen             100000 (991449 7866507, 991951 7867260, 992167 7867…
 8 NA               gen             100000 (991988 7855459, 991926 7856386, 991605 7857…
 9 NA               gen             100000 (994613 7873673, 994441 7873143, 994343 7872…
10 NA               gen             100000 (992797 7850486, 993037 7850008, 993461 7849…
# ℹ 22 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
ElveKant
}
\description{
ElveKant
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
