% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{FiktivDelelinje}
\alias{FiktivDelelinje}
\title{FiktivDelelinje}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 570 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -94248.72 ymin: 6445086 xmax: 1114719 ymax: 7953516
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 570 × 2
   oppdateringsdato                                                       grense
 * <date>                                                       <LINESTRING [m]>
 1 NA               (1e+06 7918662, 1e+06 7930421, 1e+06 7933244, 1e+06 7953516)
 2 NA                                             (1e+06 7899531, 1e+06 7918662)
 3 NA                              (1e+06 7878513, 1e+06 7892203, 1e+06 7896135)
 4 NA                              (8e+05 7877438, 8e+05 7913649, 8e+05 7921683)
 5 NA                                             (1e+06 7876833, 1e+06 7878513)
 6 NA                                             (1e+06 7874694, 1e+06 7876833)
 7 NA                                             (1e+06 7873803, 1e+06 7874694)
 8 NA                                             (1e+06 7871518, 1e+06 7873803)
 9 NA                                             (8e+05 7863049, 8e+05 7877438)
10 NA                              (1e+06 7869668, 1e+06 7870731, 1e+06 7871518)
# ℹ 560 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
FiktivDelelinje
}
\description{
FiktivDelelinje
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
