% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Fylkesgrense}
\alias{Fylkesgrense}
\title{Fylkesgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 178 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -86015.82 ymin: 6466105 xmax: 805691.2 ymax: 7853461
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 178 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000 (196838.7 7012697, 196827.5 7013311, 196731.…
 2 2023-12-18       gen             100000 (284433.7 6715571, 285547.5 6715334, 285937.…
 3 2023-12-18       gen             100000 (143290.4 6830767, 144444.4 6831800, 143735.…
 4 2023-12-18       gen             100000 (187018.9 6607767, 187010 6608128, 188507.6 …
 5 2023-12-18       gen             100000 (-1094.47 6928827, -1506.12 6925209, 4947.54…
 6 2023-12-18       gen             100000 (12564 6521784, 12798.43 6523016, 12861.48 6…
 7 2023-12-18       gen             100000 (82868.22 6686184, 82667.45 6685916, 81441.5…
 8 2023-12-18       gen             100000 (38386.01 6609345, 40247.99 6612469, 40389.9…
 9 2023-12-18       gen             100000 (40057.16 6588165, 38941.05 6588530, 37671.0…
10 2023-12-18       gen             100000 (298880.7 6708095, 299258 6707216, 299444.3 …
# ℹ 168 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Fylkesgrense
}
\description{
Fylkesgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
