% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{HavElvSperre}
\alias{HavElvSperre}
\title{HavElvSperre}
\format{
\preformatted{Simple feature collection with 3 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 265025 ymin: 6566414 xmax: 994613 ymax: 7873955
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 3 x 2
  oppdateringsdato                         geometry
* <date>                           <LINESTRING [m]>
1 2002-09-15       (994613 7873673, 993516 7873955)
2 2002-09-15       (265025 6568479, 265355 6568295)
3 2002-09-15       (268594 6566414, 269200 6566818)
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
HavElvSperre
}
\description{
HavElvSperre
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
