% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Havflate}
\alias{Havflate}
\title{Havflate}
\format{
\preformatted{Simple feature collection with 87 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99438.22 ymin: 6426073 xmax: 1121927 ymax: 7962739
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 87 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((8e+05 7859480, 799529 7859716, 799666 7860299, 8e+05 7860540, 8e~
 2 NA               ((8e+05 7803657, 8e+05 7800000, 797972.2 7800000, 797899 7800699, ~
 3 NA               ((780352.7 7800000, 778482.9 7800000, 779219 7800121, 780352.7 780~
 4 NA               ((671219.7 7800000, 672140.2 7800000, 672536 7799104, 671523 77994~
 5 NA               ((773456.3 7800000, 775949.1 7800000, 776159 7799618, 778482.9 780~
 6 NA               ((797972.2 7800000, 8e+05 7800000, 8e+05 7799360, 797882 7796456, ~
 7 NA               ((606199 7634299, 607190 7633788, 610369 7633788, 611799 7633399, ~
 8 NA               ((6e+05 7603596, 6e+05 7600000, 596165.9 7600000, 596413 7600271, ~
 9 2010-07-21       ((8e+05 7913649, 8e+05 7877438, 798699 7878999, 798399 7877899, 79~
10 NA               ((1e+06 7876833, 1e+06 7878513, 1000157 7878653, 1000726 7878688, ~
# ... with 77 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Havflate
}
\description{
Havflate
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
