% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Hoydekurve}
\alias{Hoydekurve}
\title{Hoydekurve}
\format{
\preformatted{Simple feature collection with 1507 features and 5 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -33857 ymin: 6509801 xmax: 983534 ymax: 7890593
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,507 x 6
   målemetode nøyaktighet medium oppdateringsdato høyde                        geometry
 *      <int>       <int> <chr>  <date>           <int>                <LINESTRING [m]>
 1         64      100000 T      2002-09-15         600 (560322 7646608, 559790 764615~
 2         64      100000 T      2002-09-15        1200 (651552 7646699, 651260 764667~
 3         64      100000 T      2002-09-15        1200 (678689 7645254, 677771 764517~
 4         64      100000 T      2002-09-15         600 (595017 7643368, 592588 764328~
 5         64      100000 T      2002-09-15         600 (646879 7656804, 646444 765649~
 6         64      100000 T      2002-09-15        1200 (654413 7643953, 654211 764295~
 7         64      100000 T      2002-09-15        1200 (658048 7640251, 656869 763836~
 8         64      100000 T      2002-09-15        1200 (661701 7639297, 661217 763918~
 9         64      100000 T      2002-09-15        1200 (671244 7637702, 670802 763732~
10         64      100000 T      2002-09-15         600 (589650 7638623, 590772 763933~
# ... with 1,497 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Hoyde_GML.gml}
}
\usage{
Hoydekurve
}
\description{
Høydekurve
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
