% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Hoydelag}
\alias{Hoydelag}
\title{Hoydelag}
\format{
\preformatted{Simple feature collection with 2107 features and 3 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -76204 ymin: 6450196 xmax: 1114945 ymax: 7939929
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,107 x 4
   oppdateringsdato makshøyde minhøyde                                         geometry
 * <date>               <int>    <int>                                    <POLYGON [m]>
 1 NA                     600        0 ((603103 7722650, 603861 7722540, 604655 772330~
 2 NA                    1800     1200 ((727422 7715396, 724710 7716233, 721696 771805~
 3 NA                    1200      600 ((644738 7713225, 644519 7713467, 643829 771455~
 4 NA                    1200      600 ((608324.6 7710501, 608291 7710461, 606462.3 77~
 5 NA                    1800     1200 ((722675 7707806, 721727 7708258, 721043 770893~
 6 NA                    1800     1200 ((684859.9 7704880, 684599 7704942, 684636.1 77~
 7 NA                    1200      600 ((770489 7705051, 771140 7704611, 772418 770476~
 8 NA                    1200      600 ((775091 7701701, 775762 7700632, 775878 770022~
 9 NA                    1200      600 ((659059 7696849, 658546 7696860, 657778 769712~
10 NA                    1200      600 ((636227 7695292, 635842.2 7693591, 636852 7693~
# ... with 2,097 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Hoyde_GML.gml}
}
\usage{
Hoydelag
}
\description{
Høydelag
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
