% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjo}
\alias{Innsjo}
\title{Innsjo}
\format{
\preformatted{Simple feature collection with 651 features and 2 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -47337 ymin: 6478703 xmax: 1086878 ymax: 7891559
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 651 x 3
   oppdateringsdato høyde                                                      geometry
 * <date>           <int>                                                 <POLYGON [m]>
 1 NA                  NA ((631507 7693107, 632000 7693900, 632500 7694100, 633200 769~
 2 NA                  NA ((868522.9 7887293, 868559 7887310, 868391 7885595, 868626 7~
 3 NA                  NA ((909973 7867980, 909100 7867300, 909049 7867351, 908636 786~
 4 NA                  NA ((952875 7851415, 951721 7851665, 950971 7852146, 951106 785~
 5 NA                  NA ((939380 7850092, 939194 7850143, 939600 7852600, 940000 785~
 6 NA                 207 ((945773.7 7843454, 945921.4 7842678, 946762.4 7842335, 9476~
 7 NA                  NA ((861020 7823987, 860579 7823705, 859479 7825160, 859831 782~
 8 NA                  NA ((1027488 7815542, 1027655 7816424, 1028718 7816460, 1029083~
 9 NA                  NA ((851688 7814299, 851807 7814411, 852608 7814369, 853493 781~
10 NA                  NA ((1085484 7804837, 1084800 7805100, 1082700 7809400, 1081977~
# ... with 641 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Innsjo
}
\description{
Innsjø
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
