% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjokant}
\alias{Innsjokant}
\title{Innsjokant}
\format{
\preformatted{Simple feature collection with 1277 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -47337 ymin: 6478703 xmax: 1086475 ymax: 7891559
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,277 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 2002-09-15       (1082825 7810321, 1082900 7810300, 1083593 ~
 2         64      100000 2002-09-15       (1010073 7804101, 1010866 7805223, 1011766 ~
 3         64      100000 2002-09-15       (868522.9 7887293, 868559 7887310, 868391 7~
 4         64      100000 2002-09-15       (1017831 7886223, 1017403 7888189, 1016400 ~
 5         64      100000 2002-09-15       (1022947 7881392, 1021000 7882300, 1019717 ~
 6         64      100000 2002-09-15       (913961 7871362, 914649 7872728, 914718 787~
 7         64      100000 2002-09-15       (913114 7870776, 913875 7871191, 913961 787~
 8         64      100000 2002-09-15       (913525 7871975, 911709 7871049, 910757.6 7~
 9         64      100000 2002-09-15       (909357.2 7868644, 908500 7867900, 908636 7~
10         64      100000 2002-09-15       (860500 7869800, 861300 7871000, 861600 787~
# ... with 1,267 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Innsjokant
}
\description{
Innsjøkant
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
