% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{KantUtsnitt}
\alias{KantUtsnitt}
\title{KantUtsnitt}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 16 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99438.22 ymin: 6426073 xmax: 1121927 ymax: 7962739
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 16 × 2
   oppdateringsdato                                                               grense
 * <date>                                                               <LINESTRING [m]>
 1 NA               (8e+05 7921683, 809725.8 7929806, 813250.5 7932063, 820604.2 793473…
 2 NA                                (6e+05 7777212, 606182.7 7787101, 614357.2 7800000)
 3 NA               (614357.2 7800000, 619273.9 7807759, 623331.8 7812152, 642452.2 782…
 4 NA               (431668.1 7600000, 434193.8 7604043, 438606 7611106, 447422.2 76252…
 5 NA               (4e+05 7570981, 400307 7571528, 404088.3 7576449, 409474.4 7582102,…
 6 NA               (349243.2 7400000, 351557.2 7403523, 362047.3 7414276, 354826.8 743…
 7 NA               (259619.4 7200000, 263118.3 7203317, 268151 7215831, 271850.5 72229…
 8 NA               (2e+05 7145406, 207752 7150826, 228157.8 7170172, 247755.7 7188752,…
 9 NA               (25040.09 7e+06, 38441.2 7010244, 43643.95 7014092, 48089.61 701660…
10 NA               (0 6445086, -4073.55 6447403, -4370.94 6447620, -9159.35 6451599, -…
11 NA               (0 6980291, 1439.4 6981426, 6540.3 6985445, 16385.19 6993202, 24149…
12 NA               (245118.3 6521829, 244204.9 6521982, 232007 6523843, 231755 6523881…
13 NA               (1e+06 7953516, 1007278 7950541, 1024145 7943449, 1042544 7938949, …
14 NA               (-86835.01 6800000, -87964 6812238, -86982.33 6819849, -84944.19 68…
15 NA               (-94248.72 6600000, -96325.47 6602856, -98375.07 6607301, -99438.22…
16 NA               (2e+05 6516172, 197930.6 6515309, 190403.9 6508190, 181722.6 649997…
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
KantUtsnitt
}
\description{
KantUtsnitt
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
