% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommune}
\alias{Kommune}
\title{Kommune}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 359 features and 3 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 359 × 4
   oppdateringsdato kommunenummer navn                                            område
 * <date>           <chr>         <chr>                                    <POLYGON [m]>
 1 NA               4632          Austrheim    ((-52970.05 6789050, -58675.66 6791087, …
 2 NA               4602          Kinn         ((-22368.05 6925601, -40891.32 6938004, …
 3 NA               5530          Senja        ((603053.5 7736395, 583180.8 7746329, 57…
 4 NA               1532          Giske        ((37936.65 6981824, 24150.37 6999320, 65…
 5 NA               1508          Ålesund      ((74458.6 7005817, 63609.57 7025260, 445…
 6 NA               1554          Averøy       ((115093.9 7033391, 98885.26 7048923, 92…
 7 NA               1505          Kristiansund ((123337.6 7041993, 107007.6 7057633, 98…
 8 NA               5052          Leka         ((315248.7 7258249, 295338.3 7268157, 28…
 9 NA               5618          Måsøy        ((861361.7 7923284, 852906 7943866, 8430…
10 NA               1857          Værøy        ((401855.2 7523663, 380921.3 7531107, 37…
# ℹ 349 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Kommune
}
\description{
Kommune
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
