% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommunegrense}
\alias{Kommunegrense}
\title{Kommunegrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 779 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -92079.54 ymin: 6428945 xmax: 1111593 ymax: 7962288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 779 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64      100000 2023-12-18                (283687.3 7047593, 278574.9 7048475)
 2         64      100000 2023-12-18       (597560.1 7645410, 595460 7651110, 589810 76…
 3         64      100000 2023-12-18                (174890.6 6709790, 177520.6 6706275)
 4         64      100000 2023-12-18                (278574.9 7048475, 275672.4 7051059)
 5         64      100000 2023-12-18       (34286.33 6956284, 30779.98 6957115, 27400.8…
 6         64      100000 2023-12-18                (358383.6 7358348, 365421.9 7360393)
 7         64      100000 2023-12-18                (346785.4 6772961, 349950.1 6767425)
 8         64      100000 2023-12-18       (629035.2 7657246, 631826.3 7659809, 634415.…
 9         64      100000 2023-12-18       (569570 7622020, 570609.9 7632180, 575269.8 …
10         64      100000 2023-12-18       (-34198.39 6554681, -34649.53 6554962, -3574…
# ℹ 769 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Kommunegrense
}
\description{
Kommunegrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
