% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kystkontur}
\alias{Kystkontur}
\title{Kystkontur}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2362 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -76204 ymin: 6450196 xmax: 1104122 ymax: 7939929
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,362 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64      100000 NA               (1062622 7812630, 1061905 7810437, 1061772 7…
 2         64      100000 NA               (813016 7804667, 812399 7804999, 812147 7806…
 3         64      100000 NA               (8e+05 7807639, 800237 7807513, 800609 78067…
 4         64      100000 NA               (1061071 7805289, 1060440 7805005, 1060118 7…
 5         64      100000 NA               (1080613 7803214, 1080499 7803399, 1079748 7…
 6         64      100000 NA                          (1060118 7803575, 1060021 7803145)
 7         64      100000 NA               (873003 7923569, 871124 7924428, 870252 7924…
 8         64      100000 NA               (894267 7918307, 893882 7919042, 893727 7919…
 9         64      100000 NA               (952846 7910507, 952965 7911279, 952999 7911…
10         64      100000 NA               (851199 7909000, 850599 7909799, 849299 7910…
# ℹ 2,352 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Kystkontur
}
\description{
Kystkontur
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
