% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Lufthavn}
\alias{Lufthavn}
\title{Lufthavn}
\format{
\preformatted{Simple feature collection with 50 features and 9 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -55217.49 ymin: 6472154 xmax: 1095819 ymax: 7929664
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 50 x 10
   målemetode nøyaktighet oppdateringsdato iataKode icaoKode lufthavneier lufthavntype
 *      <int>       <int> <date>           <chr>    <chr>    <chr>        <chr>       
 1         64      100000 2011-04-01       MJF      ENMS     Avinor AS    L           
 2         64      100000 2012-03-07       VRY      ENVR     Avinor AS    H           
 3         64      100000 2011-03-22       SOJ      ENSR     Avinor AS    L           
 4         64      100000 2011-03-30       FRO      ENFL     Avinor AS    L           
 5         64      100000 2011-04-04       SSJ      ENST     Avinor AS    L           
 6         64      100000 2011-03-30       FDE      ENBL     Avinor AS    L           
 7         64      100000 2011-03-29       SDN      ENSD     Avinor AS    L           
 8         64      100000 2011-03-29       BNN      ENBN     AvinorAS     L           
 9         64      100000 2011-03-31       LKN      ENLK     Avinor AS    L           
10         64      100000 2011-11-29       SVJ      ENSH     Avinor AS    L           
# ... with 40 more rows, and 3 more variables: navn <chr>, trafikktype <chr>,
#   geometry <POINT [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Lufthavn
}
\description{
Lufthavn
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
