% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Myr}
\alias{Myr}
\title{Myr}
\format{
\preformatted{Simple feature collection with 239 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -59043 ymin: 6599852 xmax: 1093877 ymax: 7912077
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 239 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((1027643 7814095, 1029126 7813766, 1029208 7814672, 1029208 78148~
 2 NA               ((116946 6645641, 116796.8 6644838, 117114.5 6643771, 116826.4 664~
 3 2010-07-21       ((891212 7685169, 891036 7683589, 889807 7683764, 889280 7683062, ~
 4 2010-07-21       ((898516 7669095, 898757 7668374, 897675 7666691, 897314 7667653, ~
 5 2010-07-21       ((896232 7677511, 897073 7676790, 897073 7675588, 897915 7675227, ~
 6 2010-07-21       ((890326.7 7670251, 889904.7 7669126, 890045.4 7668305, 890561.2 7~
 7 2010-07-21       ((862207 7686048, 862808 7684485, 863409 7684845, 864251 7685807, ~
 8 NA               ((1093877 7885671, 1091678 7886393, 1090770 7887615, 1090435 78885~
 9 NA               ((-56852.15 6600000, -57247 6599852, -57493.58 6600000, -58145 660~
10 NA               ((924020 7909883, 923422 7908986, 922824 7909385, 922036.8 7908053~
# ... with 229 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Myr
}
\description{
Myr
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
