% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N2000-package.R
\docType{package}
\name{N2000-package}
\alias{N2000}
\alias{N2000-package}
\title{N2000: Kartdata}
\source{
\itemize{
\item \url{https://kartkatalog.geonorge.no/metadata/kartverket/n2000-kartdata/8d52075d-6ef6-4120-8041-d0c0901f21f7}
}
}
\description{
Kartdata fra \href{https://kartverket.no/}{Kartverket}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hmalmedal/N2000}
  \item Report bugs at \url{https://github.com/hmalmedal/N2000/issues}
}

}
\author{
\strong{Maintainer}: Håkon Malmedal \email{hmalmedal@gmail.com}

Other contributors:
\itemize{
  \item Kartverket [copyright holder]
}

}
\keyword{package}
