% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturverngrense}
\alias{Naturverngrense}
\title{Naturverngrense}
\format{
\preformatted{Simple feature collection with 48 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 48 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 2014-01-28       (101764.9 6886328, 101142.2 6890067, 102495~
 2         64      100000 2014-01-28       (99744.91 6881318, 101377.2 6881619, 101910~
 3         64      100000 2014-01-28       (101980.8 6893894, 102246.3 6895059, 102258~
 4         64      100000 2014-01-28       (272143.5 6946763, 272789.4 6946271, 273346~
 5         64      100000 2014-01-28       (28219.27 6691623, 25983.92 6690898, 25870.~
 6         64      100000 2014-01-28       (494981.1 7472050, 493139 7471776, 493107.7~
 7         64      100000 2014-01-28       (642141.4 7608935, 638731.5 7611581, 636996~
 8         64      100000 2014-01-28       (92235.33 6740841, 92009.31 6742561, 90324.~
 9         64      100000 2014-01-28       (585497 7677287, 585690 7678220, 585370 767~
10         64      100000 2014-01-28       (517177.9 7412742, 518503 7416640, 519669 7~
# ... with 38 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Restriksjonsomrader_GML.gml}
}
\usage{
Naturverngrense
}
\description{
Naturverngrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
