% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturvernomrade}
\alias{Naturvernomrade}
\title{Naturvernomrade}
\format{
\preformatted{Simple feature collection with 42 features and 4 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 42 x 5
   oppdateringsdato navn     vernedato  verneform                              geometry
 * <date>           <chr>    <date>     <chr>                             <POLYGON [m]>
 1 2014-01-28       Breheim~ 2009-08-07 NP        ((119696.1 6895723, 120523.1 6895754~
 2 2014-01-28       Breheim~ 2009-08-07 NP        ((101980.8 6893894, 102246.3 6895059~
 3 2014-01-28       Forollh~ 2001-12-21 NP        ((270385.5 6942384, 271351.6 6946091~
 4 2014-01-28       Sjunkha~ 2010-02-05 NP        ((494981.1 7472050, 493139 7471776, ~
 5 2014-01-28       Rohkunb~ 2011-02-25 NP        ((642141.4 7608935, 638731.5 7611581~
 6 2014-01-28       Halling~ 2006-12-22 NP        ((92235.33 6740841, 92009.31 6742561~
 7 2014-01-28       Junkerd~ 2004-01-09 NP        ((517177.9 7412742, 518503 7416640, ~
 8 2014-01-28       Børgefj~ 2003-08-29 NP        ((423600 7217550, 423900 7218700, 42~
 9 2014-01-28       Gutulia  1968-12-20 NP        ((350433 6880466, 349506.3 6880776, ~
10 2021-02-09       Láhko    2012-12-14 NP        ((466777 7405812, 466713.5 7405343, ~
# ... with 32 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Restriksjonsomrader_GML.gml}
}
\usage{
Naturvernomrade
}
\description{
Naturvernområde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
