% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Riksgrense}
\alias{Riksgrense}
\title{Riksgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 40 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 245176.8 ymin: 6521821 xmax: 1121941 ymax: 7867303
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 40 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000 (560020.2 7491595, 563580.1 7498418, 566043.…
 2 2023-12-18       gen             100000 (557361.9 7461207, 554436.2 7466465, 550661.…
 3 2023-12-18       gen             100000 (322924.8 6637904, 324260.8 6638607, 325704.…
 4 2023-12-18       gen             100000 (368189.6 6712383, 367390.9 6715204, 366730.…
 5 2023-12-18       gen             100000 (478693.8 7330642, 478160.3 7334772, 482426.…
 6 2023-12-18       gen             100000 (358604.8 7054919, 362625.1 7058862, 366394.…
 7 2023-12-18       gen             100000 (468357.6 7221876, 468808.9 7224430, 469832.…
 8 2023-12-18       gen             100000 (296708.5 6544230, 296393.9 6543911, 296400.…
 9 2023-12-18       gen             100000 (931415.7 7785646, 931769.7 7785963, 932947 …
10 2023-12-18       gen             100000 (319002.8 6577574, 319087.6 6579087, 318962.…
# ℹ 30 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Riksgrense
}
\description{
Riksgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
