% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Skog}
\alias{Skog}
\title{Skog}
\format{
\preformatted{Simple feature collection with 625 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -51954 ymin: 6452847 xmax: 1114945 ymax: 7913573
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 625 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((826941.7 7800000, 824043.1 7800000, 823817.7 7800000, 817088.2 7~
 2 NA               ((802426.7 7800000, 802230.2 7800000, 800994.3 7800000, 800492.6 7~
 3 NA               ((1057671 7799864, 1057544 7799891, 1057244 7799252, 1056414 77990~
 4 NA               ((906125 7782260, 905344 7782536, 904644 7781223, 903856 7783323, ~
 5 NA               ((467910 7446094, 466445 7443630, 466105 7443373, 465640 7443021, ~
 6 NA               ((428559 7408030, 428695 7407559, 430453 7407508, 430408 7406221, ~
 7 NA               ((429699 7402399, 430599 7401899, 431899 7402099, 432643 7400944, ~
 8 NA               ((523834 7398107, 523473 7397566, 521669 7398107, 520407 7396845, ~
 9 NA               ((432680 7381576, 432255 7380469, 430979.5 7379957, 429405 7379326~
10 NA               ((413929 7353163, 415153 7353291, 417864 7352846, 417880 7352804, ~
# ... with 615 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Skog
}
\description{
Skog
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
