% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{SnoIsbre}
\alias{SnoIsbre}
\title{SnoIsbre}
\format{
\preformatted{Simple feature collection with 32 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 2502.06 ymin: 6675558 xmax: 771624 ymax: 7804288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 32 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 2009-05-05       ((699010 7743274, 699575 7742520, 700831.4 7742897, 701459 7743086~
 2 2009-05-05       ((482096.8 7425452, 482558.7 7424738, 484364.3 7424570, 483944.4 7~
 3 2009-05-05       ((140751 6867273, 141250.1 6866869, 142089.9 6865441, 138843 68638~
 4 2011-04-28       ((18242.74 6883001, 18176.15 6883001, 18093.46 6883061, 17291.97 6~
 5 2011-04-28       ((6440.68 6883563, 8369.8 6883081, 9334.36 6883081, 9897.02 688211~
 6 NA               ((616026 7575345, 618092 7573142, 618780 7573555, 620295 7573004, ~
 7 NA               ((550259 7464630, 550979 7463551, 552418 7464270, 553317 7463371, ~
 8 NA               ((477870 7403104, 477870 7401470, 477325 7400381, 476848.8 7400000~
 9 NA               ((67326 6876775, 66370 6874863, 64766.01 6873437, 64649 6873333, 6~
10 NA               ((56235 6840634, 55820.6 6840530, 55265.27 6840392, 54706 6840252,~
# ... with 22 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
SnoIsbre
}
\description{
SnøIsbre
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
