% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{StedsnavnTekst}
\alias{StedsnavnTekst}
\title{StedsnavnTekst}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2053 features and 18 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -113650.4 ymin: 6448026 xmax: 1113274 ymax: 7942102
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,053 × 19
   streng  oppdateringsdato datafangstdato tekstReferansePunktN…¹ tekstReferansePunktØst
 * <chr>   <date>           <date>                          <int>                  <int>
 1 Solør   2021-11-11       NA                                  0                      1
 2 Hadela… 2021-11-11       NA                                  0                      2
 3 Land    2021-11-11       NA                                  0                      2
 4 Romeri… 2021-11-11       NA                                  0                      0
 5 Utne    2021-11-11       NA                                  0                      1
 6 Måbøda… 2021-11-11       NA                                  0                      1
 7 Veig    2021-11-11       NA                                  0                      0
 8 Kinsar… 2021-11-11       NA                                  0                      0
 9 Jonsva… 2021-11-11       NA                                  0                      0
10 Fjergen 2021-11-11       NA                                  0                      0
# ℹ 2,043 more rows
# ℹ abbreviated name: ¹​tekstReferansePunktNord
# ℹ 14 more variables: tegnavstand <int>, skriftkode <int>, skrifttype <chr>,
#   referansemålestokk <int>, fulltekst <chr>, navneobjekthovedgruppe <int>,
#   navneobjektgruppe <int>, navneobjekttype <int>, språk <chr>,
#   stedsnavnsnummer <int>, språkprioritering <chr>, skrivemåtenummer <int>,
#   stedsnummer <int>, geometri <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Stedsnavn_GML.gml}
}
\usage{
StedsnavnTekst
}
\description{
StedsnavnTekst
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
