% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Terrengpunkt}
\alias{Terrengpunkt}
\title{Terrengpunkt}
\format{
\preformatted{Simple feature collection with 152 features and 5 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -2948 ymin: 6531310 xmax: 1105697 ymax: 7901514
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 152 x 6
   målemetode nøyaktighet medium oppdateringsdato høyde          geometry
 *      <int>       <int> <chr>  <date>           <int>       <POINT [m]>
 1         64      100000 T      2002-09-15         474 (1007557 7901514)
 2         64      100000 T      2002-09-15         481 (1045873 7895568)
 3         64      100000 T      2002-09-15         725  (997830 7883080)
 4         64      100000 T      2002-09-15         521  (920170 7870051)
 5         64      100000 T      2002-09-15         634  (868065 7865143)
 6         64      100000 T      2002-09-15         545 (1055448 7860040)
 7         64      100000 T      2002-09-15         400  (913881 7851137)
 8         64      100000 T      2002-09-15         985  (808276 7834335)
 9         64      100000 T      2002-09-15         603  (942777 7824719)
10         64      100000 T      2002-09-15        1067  (927620 7805438)
# ... with 142 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Hoyde_GML.gml}
}
\usage{
Terrengpunkt
}
\description{
Terrengpunkt
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
