% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Territorialgrense}
\alias{Territorialgrense}
\title{Territorialgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 78 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 78 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000 (417975 7588030, 428888.3 7595550, 434193.8 …
 2 2023-12-18       gen             100000 (927892.2 7959494, 951360.7 7962151, 952358.…
 3 2023-12-18       gen             100000 (-76213.86 6578577, -79111.63 6584071, -8804…
 4 2023-12-18       gen             100000          (295338.3 7268157, 296959.8 7271567)
 5 2023-12-18       gen             100000 (-86015.82 6648235, -80327.37 6659428, -7969…
 6 2023-12-18       gen             100000 (21146.36 6433669, 20724.18 6433811, 19781.6…
 7 2023-12-18       gen             100000           (63609.57 7025260, 72502.8 7030219)
 8 2023-12-18       gen             100000 (764795.7 7891994, 770308.5 7896821, 770685.…
 9 2023-12-18       gen             100000 (245176.8 6521821, 244204.9 6521982, 231754.…
10 2023-12-18       gen             100000 (-79575.13 6709024, -79515.77 6728061, -8444…
# ℹ 68 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Territorialgrense
}
\description{
Territorialgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
