% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Tettbebyggelse}
\alias{Tettbebyggelse}
\title{Tettbebyggelse}
\format{
\preformatted{Simple feature collection with 886 features and 3 fields
Geometry type: GEOMETRY
Dimension:     XY
Bounding box:  xmin: -58767 ymin: 6455265 xmax: 1105065 ymax: 7937103
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 886 x 4
   oppdateringsdato målemetode nøyaktighet                                     geometry
 * <date>                <int>       <int>                           <MULTIPOLYGON [m]>
 1 2009-05-06               NA          NA (((287507 6750012, 288432 6747854, 290236 6~
 2 2011-03-23               NA          NA (((168801 6965567, 170034 6965887, 169657 6~
 3 NA                       NA          NA (((651783 7737587, 650765 7737081, 650414.6~
 4 NA                       NA          NA (((475827 7465979, 475882.9 7465783, 476161~
 5 NA                       NA          NA (((331609 7151102, 332584 7151487, 332788 7~
 6 NA                       NA          NA (((105981 6980952, 105254 6980798, 104949.3~
 7 NA                       NA          NA (((44136 6956284, 43478 6956432, 42240 6956~
 8 NA                       NA          NA (((255261 6780785, 255400 6781400, 255500 6~
 9 NA                       NA          NA (((-42509 6733371, -41852 6733734, -42121 6~
10 NA                       NA          NA (((-42709 6732572, -41588 6730971, -41964 6~
# ... with 876 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Tettbebyggelse
}
\description{
Tettbebyggelse
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
