% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Veglenke}
\alias{Veglenke}
\title{Veglenke}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2585 features and 9 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -74177.12 ymin: 6455017 xmax: 1108413 ymax: 7937309
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,585 × 10
   medium oppdateringsdato målemetode nøyaktighet typeVeg  vegkategori vegfase vegnummer
 * <chr>  <date>           <chr>            <int> <chr>    <chr>       <chr>       <int>
 1 T      NA               gen             100000 enkelBi… E           V              39
 2 T      NA               gen             100000 enkelBi… E           V               6
 3 T      NA               gen             100000 enkelBi… E           V             134
 4 T      NA               gen             100000 enkelBi… E           V             134
 5 T      NA               gen             100000 enkelBi… E           V             134
 6 T      NA               gen             100000 enkelBi… E           V               6
 7 T      NA               gen             100000 enkelBi… E           V               6
 8 T      NA               gen             100000 enkelBi… E           V             134
 9 T      NA               gen             100000 enkelBi… F           V              NA
10 T      NA               gen             100000 enkelBi… E           V              18
# ℹ 2,575 more rows
# ℹ 2 more variables: motorvegtype <chr>, senterlinje <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Samferdsel_GML.gml}
}
\usage{
Veglenke
}
\description{
Veglenke
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
