% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{VegSenterlinje}
\alias{VegSenterlinje}
\title{VegSenterlinje}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2435 features and 8 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -59361.73 ymin: 6455017 xmax: 1108413 ymax: 7937309
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,435 × 9
   målemetode nøyaktighet medium oppdateringsdato vegkategori vegstatus vegnummer
 *      <int>       <int> <chr>  <date>           <chr>       <chr>         <int>
 1         64      100000 T      NA               E           V                39
 2         64      100000 T      NA               E           V                 6
 3         64      100000 T      NA               E           V               134
 4         64      100000 T      NA               E           V               134
 5         64      100000 T      NA               E           V               134
 6         64      100000 T      NA               E           V                 6
 7         64      100000 T      NA               E           V                 6
 8         64      100000 T      NA               E           V               134
 9         64      100000 T      NA               F           V                NA
10         64      100000 T      NA               E           V                18
# ℹ 2,425 more rows
# ℹ 2 more variables: motorvegtype <chr>, senterlinje <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Samferdsel_GML.gml}
}
\usage{
VegSenterlinje
}
\description{
VegSenterlinje
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
