% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{AnnenBatrute}
\alias{AnnenBatrute}
\title{AnnenBatrute}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 43 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -57722.3 ymin: 6734803 xmax: 1102774 ymax: 7941176
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 43 × 4
   oppdateringsdato målemetode nøyaktighet                                   senterlinje
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 NA               gen             100000 (730420.3 7782282, 730409.1 7782388, 730355.…
 2 NA               gen             100000 (393161 7324670, 392827.4 7325024, 392338 73…
 3 NA               gen             100000 (563521.5 7633773, 564021.6 7634360, 564415.…
 4 NA               gen             100000 (499837.6 7606787, 499846.4 7606783, 500613 …
 5 NA               gen             100000 (962099.2 7934023, 962724.3 7935269, 963110.…
 6 NA               gen             100000 (1055705 7841269, 1055189 7841447, 1054653 7…
 7 NA               gen             100000 (776682.3 7808699, 776682 7808701, 776590.8 …
 8 NA               gen             100000 (232374.7 7067373, 230686.7 7066181, 228555.…
 9 NA               gen             100000 (481684.6 7567549, 479999.8 7564808, 478503.…
10 NA               gen             100000 (473630.5 7463267, 472999.8 7463386, 472375.…
# ℹ 33 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Samferdsel_GML.gml}
}
\usage{
AnnenBatrute
}
\description{
AnnenBåtrute
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
