% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ApentOmrade}
\alias{ApentOmrade}
\title{ApentOmrade}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 1859 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -76204 ymin: 6450196 xmax: 1113444 ymax: 7939929
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,859 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 NA               ((282466 6754595, 282008 6757115, 282231.8 6759578, 282237 6759635,…
 2 NA               ((-31117.89 6706271, -32037.04 6705741, -32425.91 6704575, -31153.2…
 3 NA               ((112509.5 6648336, 111972.5 6648279, 111104 6648206, 110103.1 6648…
 4 NA               ((858210 7685189, 857829.2 7685522, 856444 7685696, 855391 7683940,…
 5 NA               ((307212.8 6785089, 307693 6784962, 308249.1 6785291, 308831.6 6785…
 6 NA               ((865903 7688513, 865598 7688920, 865632 7690612, 865501.1 7690472,…
 7 NA               ((644489 7719386, 645880 7719784, 646287 7720291, 645702 7720613, 6…
 8 NA               ((496899 7632004, 497818 7632386, 498117 7633337, 497301 7633361, 4…
 9 NA               ((740941 7789135, 741587 7789316, 741315 7790623, 740533 7791457, 7…
10 NA               ((892381.8 7666943, 892366.7 7667109, 892561 7667653, 892601.1 7667…
# ℹ 1,849 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
ApentOmrade
}
\description{
ÅpentOmråde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
