% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bane}
\alias{Bane}
\title{Bane}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 109 features and 4 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -42446 ymin: 6467440 xmax: 1076891 ymax: 7806691
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 109 × 5
   medium oppdateringsdato målemetode nøyaktighet                            senterlinje
 * <chr>  <date>           <chr>            <int>                       <LINESTRING [m]>
 1 U      NA               gen             100000 (235683 6634536, 235844.6 6634643, 23…
 2 T      NA               gen             100000 (285556.1 6670637, 285902.2 6671698, …
 3 U      NA               gen             100000 (-30553 6732995, -30335.87 6732941, -…
 4 T      NA               gen             100000 (267168.3 6650542, 267311 6650895, 26…
 5 T      NA               gen             100000 (371748 7150284, 371962.4 7149771, 37…
 6 T      NA               gen             100000 (255598 6650050, 255842.8 6649937, 25…
 7 T      NA               gen             100000 (244465.7 6642784, 244817.9 6643195, …
 8 T      NA               gen             100000   (248813.7 6647682, 249009.9 6647662)
 9 T      NA               gen             100000 (256066.1 6649833, 256241.7 6649941, …
10 T      NA               gen             100000 (187931.3 6648911, 188304.3 6648674, …
# ℹ 99 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Samferdsel_GML.gml}
}
\usage{
Bane
}
\description{
Bane
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
