% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bilferjestrekning}
\alias{Bilferjestrekning}
\title{Bilferjestrekning}
\format{
\preformatted{Simple feature collection with 154 features and 6 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -74177.12 ymin: 6522919 xmax: 849429.5 ymax: 7904527
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 154 x 7
   målemetode nøyaktighet medium oppdateringsdato vegkategori vegnummer
 *      <int>       <int> <chr>  <date>           <chr>           <int>
 1         64      100000 T      2010-07-06       F                 668
 2         64      100000 T      2010-07-06       F                 659
 3         64      100000 T      2010-07-06       F                 659
 4         64      100000 T      2010-07-06       F                 659
 5         64      100000 T      2008-10-16       E                  39
 6         64      100000 T      2010-07-06       F                 808
 7         64      100000 T      2010-07-06       F                  17
 8         64      100000 T      2010-07-06       F                 769
 9         64      100000 T      2010-07-06       F                 715
10         64      100000 T      2010-07-06       F                 670
# ... with 144 more rows, and 1 more variable: geometry <LINESTRING [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Samferdsel_GML.gml}
}
\usage{
Bilferjestrekning
}
\description{
Bilferjestrekning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
