% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ElvBekk}
\alias{ElvBekk}
\title{ElvBekk}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2393 features and 4 fields
Geometry type: GEOMETRY
Dimension:     XY
Bounding box:  xmin: -47954 ymin: 6455920 xmax: 1113674 ymax: 7935049
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,393 × 5
   oppdateringsdato vannBredde målemetode nøyaktighet                           geometry
 * <date>                <int>      <int>       <int>                     <GEOMETRY [m]>
 1 NA                        7         NA          NA POLYGON ((287769 6604306, 287136 …
 2 NA                        7         NA          NA POLYGON ((994613 7873673, 993516 …
 3 NA                        7         NA          NA POLYGON ((269200 6566818, 269044 …
 4 NA                        4         64      100000 LINESTRING (468155 7411803, 46799…
 5 NA                        4         64      100000 LINESTRING (516944 7410751, 51772…
 6 NA                        4         64      100000 LINESTRING (474184 7409740, 47482…
 7 NA                        4         64      100000 LINESTRING (529357 7405602, 52849…
 8 NA                        4         64      100000 LINESTRING (489827 7396233, 48916…
 9 NA                        4         64      100000 LINESTRING (516600 7388341, 51673…
10 NA                        4         64      100000 LINESTRING (498640 7376769, 49862…
# ℹ 2,383 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
ElvBekk
}
\description{
ElvBekk
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
