% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ElvBekk}
\alias{ElvBekk}
\title{ElvBekk}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2390 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -47954 ymin: 6455920 xmax: 1113674 ymax: 7935049
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,390 × 2
   vannbredde                                                                senterlinje
 *      <int>                                                           <LINESTRING [m]>
 1          4                                           (468155 7411803, 467994 7411624)
 2          4                                           (516944 7410751, 517726 7409785)
 3          4 (474184 7409740, 474824 7409968, 475228 7410026, 475656 7409891, 476542 7…
 4          4 (529357 7405602, 528492 7405996, 526588 7407765, 525426 7408936, 524399 7…
 5          4 (489827 7396233, 489168 7396698, 488947 7396950, 488864 7397242, 488949 7…
 6          4 (516600 7388341, 516737 7389133, 518290.6 7394354, 518370 7394621, 518603…
 7          4 (498640 7376769, 498624 7376761, 498030.7 7376462, 497217 7376051, 496964…
 8          4 (486242 7392039, 484731 7388515, 484410 7386980, 484300 7386417, 484720.5…
 9          4 (505913 7369806, 504845 7370505, 504400.7 7371243, 504080 7371776, 501248…
10          4 (495500 7375997, 494516 7375185, 493519.4 7372395, 493382 7372010, 493363…
# ℹ 2,380 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
ElvBekk
}
\description{
ElvBekk
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
