% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Fylkesgrense}
\alias{Fylkesgrense}
\title{Fylkesgrense}
\format{
\preformatted{Simple feature collection with 158 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -86015.82 ymin: 6466105 xmax: 627600.6 ymax: 7721400
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 158 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 2020-12-14       (322924.8 6637904, 322662.6 6638501, 322634~
 2         64      100000 2020-12-14       (56959.6 6653689, 57492.86 6651438, 57166.6~
 3         64      100000 2020-12-14       (623019.9 7613727, 623406.8 7613041, 624087~
 4         64      100000 2020-12-14       (613885.9 7612332, 617030.9 7612871, 619380~
 5         64      100000 2020-12-14       (603726.5 7611575, 605570 7611671, 607932.8~
 6         64      100000 2020-12-14               (595331.1 7607386, 603726.5 7611575)
 7         64      100000 2020-12-14       (564300 7592100, 563843.1 7593529, 563564.9~
 8         64      100000 2020-12-14       (542350 7585230, 547800 7584800, 552890 758~
 9         64      100000 2020-12-14       (541750 7607552, 541849 7607391, 541876 760~
10         64      100000 2020-12-14       (525157.5 7603548, 525714.1 7602849, 526672~
# ... with 148 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Fylkesgrense
}
\description{
Fylkesgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
