% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Fylkesgrense}
\alias{Fylkesgrense}
\title{Fylkesgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 178 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -86015.82 ymin: 6466105 xmax: 805691.2 ymax: 7853461
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 178 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64      100000 2023-12-18                    (249465 6576412, 248748 6578266)
 2         64      100000 2023-12-18       (193105.8 6929798, 194831.8 6928574, 200583.…
 3         64      100000 2023-12-18       (255714.8 6674207, 255667.1 6674670, 254751.…
 4         64      100000 2023-12-18                (272293.1 6639650, 272134.9 6640209)
 5         64      100000 2023-12-18       (242698.7 6602286, 239453.5 6606398, 239819.…
 6         64      100000 2023-12-18       (98029.68 6719704, 95555.22 6706697, 91406.9…
 7         64      100000 2023-12-18                (323077.7 6661471, 322764.8 6658306)
 8         64      100000 2023-12-18       (76948.77 6608219, 76969.77 6608285, 77037.7…
 9         64      100000 2023-12-18       (190047.6 6742680, 190003.2 6741999, 189971.…
10         64      100000 2023-12-18       (437120.3 7221190, 439323.3 7221592, 440547.…
# ℹ 168 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Fylkesgrense
}
\description{
Fylkesgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
