% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Fylkesgrense}
\alias{Fylkesgrense}
\title{Fylkesgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 178 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -86015.82 ymin: 6466105 xmax: 805691.2 ymax: 7853461
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 178 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000 (726348.8 7803538, 711663.9 7833510, 708388.…
 2 2023-12-18       gen             100000 (176112.8 6931885, 175400.6 6931909, 170490.…
 3 2023-12-18       gen             100000 (240763.5 6650012, 241360.9 6650400, 241600.…
 4 2023-12-18       gen             100000 (266264.1 6613296, 265681.2 6612583, 265503.…
 5 2023-12-18       gen             100000          (215836.7 6713407, 215837.9 6713404)
 6 2023-12-18       gen             100000          (248135.2 6663613, 248657.3 6661910)
 7 2023-12-18       gen             100000 (284648.3 6956864, 286399.3 6957780, 287760.…
 8 2023-12-18       gen             100000 (199512.8 6552951, 199412.9 6551955, 199493.…
 9 2023-12-18       gen             100000 (242698.7 6602286, 239453.5 6606398, 239819.…
10 2023-12-18       gen             100000 (98029.68 6719704, 95555.22 6706697, 91406.9…
# ℹ 168 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Fylkesgrense
}
\description{
Fylkesgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
