% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Havflate}
\alias{Havflate}
\title{Havflate}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 87 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99438.22 ymin: 6426073 xmax: 1121927 ymax: 7962739
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 87 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 NA               ((671219.7 7800000, 671523 7799402, 672536 7799104, 672140.2 780000…
 2 NA               ((0 6611322, 0 6610043, 84 6610222, 903 6610549, 1175 6611059, 357 …
 3 NA               ((8e+05 7859480, 8e+05 7860540, 799666 7860299, 799529 7859716, 8e+…
 4 NA               ((8e+05 7803657, 799865 7803082, 798499 7803899, 799122 7802181, 79…
 5 NA               ((780352.7 7800000, 779219 7800121, 778482.9 7800000, 780352.7 7800…
 6 NA               ((773456.3 7800000, 773300 7799260, 774276 7799403, 774323 7798701,…
 7 NA               ((797972.2 7800000, 797974 7799983, 796579 7798250, 795989 7797936,…
 8 NA               ((606199 7634299, 605005.4 7635043, 602089 7636861, 6e+05 7637409, …
 9 NA               ((6e+05 7603596, 598953 7603237, 596413 7600271, 596165.9 7600000, …
10 NA               ((8e+05 7913649, 8e+05 7921683, 796042.7 7918378, 782924.6 7907393,…
# ℹ 77 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Havflate
}
\description{
Havflate
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
