% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Hoydekurve}
\alias{Hoydekurve}
\title{Hoydekurve}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 1507 features and 5 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -33857 ymin: 6509801 xmax: 983534 ymax: 7890593
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,507 × 6
   medium oppdateringsdato målemetode nøyaktighet høyde                      senterlinje
 * <chr>  <date>           <chr>            <int> <int>                 <LINESTRING [m]>
 1 T      NA               gen             100000   600 (560322 7646608, 559790 7646158…
 2 T      NA               gen             100000  1200 (651552 7646699, 651260 7646671…
 3 T      NA               gen             100000  1200 (678689 7645254, 677771 7645175…
 4 T      NA               gen             100000   600 (595017 7643368, 592588 7643284…
 5 T      NA               gen             100000   600 (646879 7656804, 646444 7656499…
 6 T      NA               gen             100000  1200 (654413 7643953, 654211 7642956…
 7 T      NA               gen             100000  1200 (658048 7640251, 656869 7638362…
 8 T      NA               gen             100000  1200 (661701 7639297, 661217 7639186…
 9 T      NA               gen             100000  1200 (671244 7637702, 670802 7637327…
10 T      NA               gen             100000   600 (589650 7638623, 590772 7639339…
# ℹ 1,497 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Hoyde_GML.gml}
}
\usage{
Hoydekurve
}
\description{
Høydekurve
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
