% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Hoydelag}
\alias{Hoydelag}
\title{Hoydelag}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2107 features and 3 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -76204 ymin: 6450196 xmax: 1114945 ymax: 7939929
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,107 × 4
   oppdateringsdato makshøyde minhøyde                                            område
 * <date>               <int>    <int>                                     <POLYGON [m]>
 1 NA                     600        0 ((603103 7722650, 602069 7723385, 602827 7721952…
 2 NA                    1800     1200 ((727422 7715396, 727557 7715670, 727157 7718011…
 3 NA                    1200      600 ((644738 7713225, 645154 7713300, 645986 7714578…
 4 NA                    1200      600 ((608324.6 7710501, 608579 7710805, 608688 77111…
 5 NA                    1800     1200 ((722675 7707806, 723918 7708013, 725093 7708327…
 6 NA                    1800     1200 ((684859.9 7704880, 685862.8 7706122, 686713 770…
 7 NA                    1200      600 ((770489 7705051, 769016 7704992, 768897 7704949…
 8 NA                    1200      600 ((775091 7701701, 774967 7701746, 773971.9 77016…
 9 NA                    1200      600 ((659059 7696849, 659281 7697851, 659285 7698340…
10 NA                    1200      600 ((636227 7695292, 636107 7695613, 635869 7695873…
# ℹ 2,097 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Hoyde_GML.gml}
}
\usage{
Hoydelag
}
\description{
Høydelag
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
