% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjo}
\alias{Innsjo}
\title{Innsjo}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 651 features and 2 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -47337 ymin: 6478703 xmax: 1086878 ymax: 7891559
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 651 × 3
   oppdateringsdato høyde                                                         område
 * <date>           <int>                                                  <POLYGON [m]>
 1 NA                  NA ((92600 6541000, 91800 6540800, 91300 6540300, 93600 6538300,…
 2 NA                  NA ((631507 7693107, 631853 7692075, 632301 7691155, 632923 7691…
 3 NA                  NA ((868522.9 7887293, 867700 7886900, 867400 7885800, 866448 78…
 4 NA                  NA ((909973 7867980, 910344 7868268, 911089.2 7868848, 911176 78…
 5 NA                  NA ((952875 7851415, 952913 7851645, 952163 7852722, 952163 7852…
 6 NA                  NA ((939380 7850092, 939625 7850026, 940901 7849679, 941200 7849…
 7 NA                 207 ((945773.7 7843454, 945305.5 7842694, 945451.8 7841672, 94476…
 8 NA                  NA ((861020 7823987, 861283 7824155, 861987 7823967, 863184 7823…
 9 NA                  NA ((1027488 7815542, 1028585 7814971, 1029100 7815700, 1029208 …
10 NA                  NA ((851688 7814299, 850300 7813000, 850600 7812500, 851400 7813…
# ℹ 641 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Innsjo
}
\description{
Innsjø
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
