% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjokant}
\alias{Innsjokant}
\title{Innsjokant}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 1277 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -47337 ymin: 6478703 xmax: 1086475 ymax: 7891559
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,277 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 NA               gen             100000 (1082825 7810321, 1082900 7810300, 1083593 7…
 2 NA               gen             100000 (1010073 7804101, 1010866 7805223, 1011766 7…
 3 NA               gen             100000 (868522.9 7887293, 868559 7887310, 868391 78…
 4 NA               gen             100000 (1017831 7886223, 1017403 7888189, 1016400 7…
 5 NA               gen             100000 (1022947 7881392, 1021000 7882300, 1019717 7…
 6 NA               gen             100000 (913961 7871362, 914649 7872728, 914718 7873…
 7 NA               gen             100000 (913114 7870776, 913875 7871191, 913961 7871…
 8 NA               gen             100000 (913525 7871975, 911709 7871049, 910757.6 78…
 9 NA               gen             100000 (909357.2 7868644, 908500 7867900, 908636 78…
10 NA               gen             100000 (860500 7869800, 861300 7871000, 861600 7871…
# ℹ 1,267 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Innsjokant
}
\description{
Innsjøkant
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
