% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommune}
\alias{Kommune}
\title{Kommune}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 359 features and 5 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 359 × 6
   oppdateringsdato kommunenummer kommunenavn fylkesnummer fylkesnavn
 * <date>           <chr>         <chr>              <int> <chr>     
 1 2023-12-18       3436          Nord-Fron             34 Innlandet 
 2 2023-12-18       3415          Sør-Odal              34 Innlandet 
 3 2023-12-18       5014          Frøya                 50 Trøndelag 
 4 2023-12-18       4645          Askvoll               46 Vestland  
 5 2023-12-18       3330          Hol                   33 Buskerud  
 6 2023-12-18       3425          Engerdal              34 Innlandet 
 7 2023-12-18       1828          Nesna                 18 Nordland  
 8 2023-12-18       5603          Hammerfest            56 Finnmark  
 9 2023-12-18       1114          Bjerkreim             11 Rogaland  
10 2023-12-18       5001          Trondheim             50 Trøndelag 
# ℹ 349 more rows
# ℹ 1 more variable: område <POLYGON [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Kommune
}
\description{
Kommune
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
