% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommune}
\alias{Kommune}
\title{Kommune}
\format{
\preformatted{Simple feature collection with 357 features and 3 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 357 x 4
   oppdateringsdato kommunenummer navn                                         geometry
 * <date>           <chr>         <chr>                                   <POLYGON [m]>
 1 2020-12-14       1101          Eigersund ((4174.82 6532535, 4894.09 6532017, 5548.8~
 2 2020-12-14       4219          Evje og ~ ((75631.54 6533055, 76399.15 6532400, 7665~
 3 2020-12-14       4201          Risør     ((164788.6 6534867, 164785.1 6534775, 1647~
 4 2020-12-14       1149          Karmøy    ((-69037.3 6632413, -63419.96 6627708, -59~
 5 2020-12-14       4212          Vegårshei ((144188.3 6541506, 145287.5 6540645, 1472~
 6 2020-12-14       1119          Hå        ((-43159.21 6544994, -42556.53 6544879, -4~
 7 2020-12-14       4227          Kvinesdal ((41743.64 6547099, 42213.8 6547092, 42918~
 8 2020-12-14       4218          Iveland   ((88984.56 6510675, 88945.27 6510355, 8894~
 9 2020-12-14       1114          Bjerkreim ((-7949.72 6550366, -6469.44 6549224, -411~
10 2020-12-14       4224          Åseral    ((65590.71 6551161, 65630.88 6551147, 6583~
# ... with 347 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Kommune
}
\description{
Kommune
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
