% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommune}
\alias{Kommune}
\title{Kommune}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 359 features and 5 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 359 × 6
   oppdateringsdato kommunenummer kommunenavn fylkesnummer fylkesnavn     
 * <date>           <chr>         <chr>       <chr>        <chr>          
 1 2023-12-18       4632          Austrheim   46           Vestland       
 2 2023-12-18       1576          Aure        15           Møre og Romsdal
 3 2023-12-18       4202          Grimstad    42           Agder          
 4 2023-12-18       5622          Porsanger   56           Finnmark       
 5 2023-12-18       1860          Vestvågøy   18           Nordland       
 6 2023-12-18       1517          Hareid      15           Møre og Romsdal
 7 2023-12-18       1547          Aukra       15           Møre og Romsdal
 8 2023-12-18       3911          Færder      39           Vestfold       
 9 2023-12-18       5034          Meråker     50           Trøndelag      
10 2023-12-18       1532          Giske       15           Møre og Romsdal
# ℹ 349 more rows
# ℹ 1 more variable: område <POLYGON [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Kommune
}
\description{
Kommune
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
