% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommunegrense}
\alias{Kommunegrense}
\title{Kommunegrense}
\format{
\preformatted{Simple feature collection with 867 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -92079.54 ymin: 6428945 xmax: 1111593 ymax: 7962288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 867 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 2020-12-14       (706270.6 7715032, 703692.9 7709249, 702008~
 2         64      100000 2020-12-14               (376029.1 7443716, 354846.2 7437006)
 3         64      100000 2020-12-14                 (384820 7447740, 376029.1 7443716)
 4         64      100000 2020-12-14       (252186.6 6607757, 253351.4 6606957, 254480~
 5         64      100000 2020-12-14       (252186.6 6607757, 253308.4 6611273, 253633~
 6         64      100000 2020-12-14       (158111.5 6752937, 158378.8 6751194, 158706~
 7         64      100000 2020-12-14       (158111.5 6752937, 158401.3 6753494, 158460~
 8         64      100000 2020-12-14       (158111.5 6752937, 158077.1 6753240, 157016~
 9         64      100000 2020-12-14       (357997.7 7175873, 353554.2 7173075, 351015~
10         64      100000 2020-12-14       (357997.7 7175873, 358397.9 7174867, 358286~
# ... with 857 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Kommunegrense
}
\description{
Kommunegrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
