% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommunegrense}
\alias{Kommunegrense}
\title{Kommunegrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 779 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -92079.54 ymin: 6428945 xmax: 1111593 ymax: 7962288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 779 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000 (64739.98 6510403, 64782.46 6510323, 64804.3…
 2 2023-12-18       gen             100000 (34031.86 6804523, 29577.21 6807670, 26785.0…
 3 2023-12-18       gen             100000 (253565.8 7045761, 256221.5 7046426, 262895.…
 4 2023-12-18       gen             100000 (-16558.74 6692132, -16614.41 6693078, -1651…
 5 2023-12-18       gen             100000 (738347 7692234, 737332.4 7693228, 736290.4 …
 6 2023-12-18       gen             100000 (-53035.57 6686173, -56523.84 6686120, -7964…
 7 2023-12-18       gen             100000 (179456.5 6890333, 179027.2 6886087, 179263.…
 8 2023-12-18       gen             100000 (50095.06 6731657, 49719.99 6732666, 49588.5…
 9 2023-12-18       gen             100000 (318128.5 6895075, 320007.5 6892676, 320481.…
10 2023-12-18       gen             100000 (137571.6 6511440, 137164.2 6511712, 136156.…
# ℹ 769 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Kommunegrense
}
\description{
Kommunegrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
