% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommunegrense}
\alias{Kommunegrense}
\title{Kommunegrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 779 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -92079.54 ymin: 6428945 xmax: 1111593 ymax: 7962288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 779 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000 (271634.6 6693537, 270495.1 6691696, 268300.…
 2 2023-12-18       gen             100000 (-53035.57 6686173, -56523.84 6686120, -7964…
 3 2023-12-18       gen             100000 (137571.6 6511440, 137164.2 6511712, 136156.…
 4 2023-12-18       gen             100000 (311265.2 6991492, 320807.4 6987739, 322581.…
 5 2023-12-18       gen             100000 (275838.4 7079511, 278735.4 7082342, 283594.…
 6 2023-12-18       gen             100000          (39544.67 6956626, 37192.82 6955954)
 7 2023-12-18       gen             100000 (-15056.21 6736981, -14528.55 6736693, -1390…
 8 2023-12-18       gen             100000 (101766 6475740, 101707.9 6476722, 102188.8 …
 9 2023-12-18       gen             100000 (66785.17 6980720, 62549.1 6975781, 64687.27…
10 2023-12-18       gen             100000 (152677.3 7018684, 148348.8 7018758, 143899.…
# ℹ 769 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Kommunegrense
}
\description{
Kommunegrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
