% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Lufthavn}
\alias{Lufthavn}
\title{Lufthavn}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 49 features and 9 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -55217.49 ymin: 6472154 xmax: 1095819 ymax: 7929664
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 49 × 10
   målemetode nøyaktighet oppdateringsdato iataKode icaoKode lufthavneier lufthavntype
 *      <int>       <int> <date>           <chr>    <chr>    <chr>        <chr>       
 1         64      100000 NA               MJF      ENMS     Avinor AS    L           
 2         64      100000 NA               VRY      ENVR     Avinor AS    H           
 3         64      100000 NA               SOJ      ENSR     Avinor AS    L           
 4         64      100000 NA               FRO      ENFL     Avinor AS    L           
 5         64      100000 NA               SSJ      ENST     Avinor AS    L           
 6         64      100000 NA               FDE      ENBL     Avinor AS    L           
 7         64      100000 NA               SDN      ENSD     Avinor AS    L           
 8         64      100000 NA               BNN      ENBN     AvinorAS     L           
 9         64      100000 NA               LKN      ENLK     Avinor AS    L           
10         64      100000 NA               SVJ      ENSH     Avinor AS    L           
# ℹ 39 more rows
# ℹ 3 more variables: navn <chr>, trafikktype <chr>, posisjon <POINT [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Lufthavn
}
\description{
Lufthavn
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
