% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Lufthavn}
\alias{Lufthavn}
\title{Lufthavn}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 49 features and 6 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -55217.49 ymin: 6472154 xmax: 1095819 ymax: 7929664
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 49 × 7
   oppdateringsdato målemetode nøyaktighet lufthavntype trafikktype navn                
 * <date>           <chr>            <int> <chr>        <chr>       <chr>               
 1 NA               gen             100000 L            N           Mosjøen lufthavn, K…
 2 NA               gen             100000 H            N           Værøy flyplass      
 3 NA               gen             100000 L            N           Sørkjosen lufthavn  
 4 NA               gen             100000 L            N           Florø lufthamn      
 5 NA               gen             100000 L            N           Sandnessjøen luftha…
 6 NA               gen             100000 L            N           Førde lufthamn, Bri…
 7 NA               gen             100000 L            N           Sandane lufthamn, A…
 8 NA               gen             100000 L            N           Brønnøysund lufthav…
 9 NA               gen             100000 L            N           Leknes lufthavn     
10 NA               gen             100000 L            N           Svolvær lufthavn, H…
# ℹ 39 more rows
# ℹ 1 more variable: posisjon <POINT [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Lufthavn
}
\description{
Lufthavn
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
