% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Myr}
\alias{Myr}
\title{Myr}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 239 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -59043 ymin: 6599852 xmax: 1093877 ymax: 7912077
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 239 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 NA               ((1027643 7814095, 1027231 7815084, 1026490 7815084, 1026371 781477…
 2 NA               ((116946 6645641, 116365 6646030, 114591 6646495, 114390 6646888, 1…
 3 NA               ((891212 7685169, 889456 7684467, 888417 7684004, 888297 7681238, 8…
 4 NA               ((898516 7669095, 897554 7669095, 893707 7668614, 893672.9 7668675,…
 5 NA               ((896232 7677511, 895270 7677151, 894308 7677391, 894042.3 7677125,…
 6 NA               ((890326.7 7670251, 890959.8 7671142, 891077.1 7672033, 890561.2 76…
 7 NA               ((862207 7686048, 861005 7684605, 860132 7684467, 859781 7683062, 8…
 8 NA               ((1093877 7885671, 1093877 7887752, 1093096 7888143, 1091315 788903…
 9 NA               ((-56852.15 6600000, -55809 6600391, -53833 6603085, -53653 6604882…
10 NA               ((1021971 7852563, 1020581 7851821, 1018541 7851728, 1016131 785098…
# ℹ 229 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Myr
}
\description{
Myr
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
