% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturvernomrade}
\alias{Naturvernomrade}
\title{Naturvernomrade}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 43 features and 4 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 43 × 5
   oppdateringsdato vernedato  verneform navn                                     område
 * <date>           <date>     <chr>     <chr>                             <POLYGON [m]>
 1 2017-01-05       2009-08-07 NP        Breheimen          ((119696.1 6895723, 119277.…
 2 2017-01-05       2009-08-07 NP        Breheimen          ((101980.8 6893894, 104218.…
 3 2017-01-05       2001-12-21 NP        Forollhogna        ((270385.5 6942384, 270875.…
 4 2017-01-05       2010-02-05 NP        Sjunkhatten        ((494981.1 7472050, 495072.…
 5 2017-01-05       2011-02-25 NP        Rohkunborri        ((642141.4 7608935, 645026.…
 6 2017-01-05       2006-12-22 NP        Hallingskarvet     ((92235.33 6740841, 92255.1…
 7 2017-01-05       2002-12-20 NP        Stabbursdalen      ((845960.9 7801580, 846028.…
 8 2017-01-05       2004-01-09 NP        Junkerdal          ((517177.9 7412742, 518334.…
 9 2017-01-05       2003-08-29 NP        Børgefjell/Byrkije ((423600 7217550, 426278.5 …
10 2017-01-05       1968-12-20 NP        Gutulia            ((350433 6880466, 350477.3 …
# ℹ 33 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Restriksjonsomrader_GML.gml}
}
\usage{
Naturvernomrade
}
\description{
Naturvernområde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
