% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Navigasjonsinstallasjon}
\alias{Navigasjonsinstallasjon}
\title{Navigasjonsinstallasjon}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 106 features and 3 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -75270 ymin: 6449852 xmax: 1098990 ymax: 7939496
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 106 × 4
   målemetode nøyaktighet oppdateringsdato          posisjon
 *      <int>       <int> <date>                 <POINT [m]>
 1         64      100000 NA                (974767 7939496)
 2         64      100000 NA               (1016912 7922064)
 3         64      100000 NA                (904432 7921858)
 4         64      100000 NA               (1050929 7913703)
 5         64      100000 NA               (1098990 7889436)
 6         64      100000 NA               (1077449 7824528)
 7         64      100000 NA                (693512 7809383)
 8         64      100000 NA                (669510 7799554)
 9         64      100000 NA                (610040 7724001)
10         64      100000 NA                (506772 7661778)
# ℹ 96 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000BygningerOgAnlegg_GML.gml}
}
\usage{
Navigasjonsinstallasjon
}
\description{
Navigasjonsinstallasjon
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
