% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Navigasjonsinstallasjon}
\alias{Navigasjonsinstallasjon}
\title{Navigasjonsinstallasjon}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 106 features and 3 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -75270 ymin: 6449852 xmax: 1098990 ymax: 7939496
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 106 × 4
   oppdateringsdato målemetode nøyaktighet          posisjon
 * <date>           <chr>            <int>       <POINT [m]>
 1 NA               gen             100000  (974767 7939496)
 2 NA               gen             100000 (1016912 7922064)
 3 NA               gen             100000  (904432 7921858)
 4 NA               gen             100000 (1050929 7913703)
 5 NA               gen             100000 (1098990 7889436)
 6 NA               gen             100000 (1077449 7824528)
 7 NA               gen             100000  (693512 7809383)
 8 NA               gen             100000  (669510 7799554)
 9 NA               gen             100000  (610040 7724001)
10 NA               gen             100000  (506772 7661778)
# ℹ 96 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000BygningerOgAnlegg_GML.gml}
}
\usage{
Navigasjonsinstallasjon
}
\description{
Navigasjonsinstallasjon
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
