% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Riksgrense}
\alias{Riksgrense}
\title{Riksgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 40 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 245176.8 ymin: 6521821 xmax: 1121941 ymax: 7867303
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 40 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64      100000 2023-12-18       (701248.5 7613149, 705613.4 7617213, 709539.…
 2         64      100000 2023-12-18       (468357.6 7221876, 468808.9 7224430, 469832.…
 3         64      100000 2023-12-18       (355351.3 6748348, 355192.2 6749286, 355251 …
 4         64      100000 2023-12-18       (358604.8 7054919, 362625.1 7058862, 366394.…
 5         64      100000 2023-12-18       (931415.7 7785646, 931769.7 7785963, 932947 …
 6         64      100000 2023-12-18       (916230.4 7677406, 916837.2 7677804, 918707.…
 7         64      100000 2023-12-18       (322924.8 6637904, 324260.8 6638607, 325704.…
 8         64      100000 2023-12-18       (747182.8 7701826, 750698 7701151, 753776.2 …
 9         64      100000 2023-12-18       (296708.5 6544230, 296393.9 6543911, 296400.…
10         64      100000 2023-12-18       (357800.3 6914076, 355563.4 6921989, 352719.…
# ℹ 30 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Riksgrense
}
\description{
Riksgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
