% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Riksgrense}
\alias{Riksgrense}
\title{Riksgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 40 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 245176.8 ymin: 6521821 xmax: 1121941 ymax: 7867303
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 40 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000            (1008758 7808534, 1008977 7804953)
 2 2023-12-18       gen             100000 (468357.6 7221876, 468808.9 7224430, 469832.…
 3 2023-12-18       gen             100000 (747182.8 7701826, 750698 7701151, 753776.2 …
 4 2023-12-18       gen             100000 (353736.6 6960754, 353016.3 6966019, 352483.…
 5 2023-12-18       gen             100000 (349950.1 6767425, 350828.8 6768044, 350980 …
 6 2023-12-18       gen             100000 (449269.1 7189059, 451534.2 7193206, 453371.…
 7 2023-12-18       gen             100000 (296708.5 6544230, 296393.9 6543911, 296400.…
 8 2023-12-18       gen             100000 (245176.8 6521821, 248777.1 6536249, 255691.…
 9 2023-12-18       gen             100000 (314938.8 6614098, 315066.2 6614272, 318070.…
10 2023-12-18       gen             100000 (355351.3 6748348, 355192.2 6749286, 355251 …
# ℹ 30 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Riksgrense
}
\description{
Riksgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
