% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Skog}
\alias{Skog}
\title{Skog}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 625 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -51954 ymin: 6452847 xmax: 1114945 ymax: 7913573
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 625 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 NA               ((826941.7 7800000, 828220 7800639, 827139 7801360, 827980 7802921,…
 2 NA               ((802426.7 7800000, 802721 7800483, 802737 7800920, 801453 7801352,…
 3 NA               ((1057671 7799864, 1057549 7799925, 1058303 7800000, 1052471 780000…
 4 NA               ((906125 7782260, 907267 7782926, 907077 7783973, 906220 7784259, 9…
 5 NA               ((467910 7446094, 467954 7446169, 467719 7447099, 467549 7447538, 4…
 6 NA               ((428559 7408030, 427077 7407981, 426141 7407349, 427319 7406045, 4…
 7 NA               ((429699 7402399, 428699 7402699, 428571.8 7402679, 426100 7402299,…
 8 NA               ((523834 7398107, 522978.7 7399676, 522802.2 7400000, 520879.8 7400…
 9 NA               ((432680 7381576, 430604 7381786, 430027 7381133, 427864 7380979, 4…
10 NA               ((413929 7353163, 414158 7351917, 410709 7349557, 412068 7351810, 4…
# ℹ 615 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Skog
}
\description{
Skog
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
