% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{SnoIsbre}
\alias{SnoIsbre}
\title{SnoIsbre}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 32 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 2502.06 ymin: 6675558 xmax: 771624 ymax: 7804288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 32 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 NA               ((699010 7743274, 698436.1 7742905, 696373 7741579, 696232.5 774165…
 2 NA               ((482096.8 7425452, 480879.1 7425578, 481047 7426291, 480291.2 7426…
 3 NA               ((140751 6867273, 139802.8 6868215, 138784.5 6867143, 136667.5 6867…
 4 NA               ((18242.74 6883001, 18979.95 6883001, 20507.16 6883001, 20828.68 68…
 5 NA               ((6440.68 6883563, 4833.08 6883322, 3979.2 6882619, 3466.62 6882197…
 6 NA               ((616026 7575345, 615062 7575759, 614787 7575345, 615062 7573968, 6…
 7 NA               ((550259 7464630, 548273.5 7465028, 547562 7465170, 547457.4 746533…
 8 NA               ((477870 7403104, 479866 7405827, 480774 7408005, 479685 7408913, 4…
 9 NA               ((67326 6876775, 67071.28 6878178, 66944 6878879, 65414 6880217, 64…
10 NA               ((56235 6840634, 57306.37 6843490, 57383 6843694, 57128.88 6844202,…
# ℹ 22 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
SnoIsbre
}
\description{
SnøIsbre
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
