% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Stasjon}
\alias{Stasjon}
\title{Stasjon}
\format{
\preformatted{Simple feature collection with 94 features and 4 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -40392.89 ymin: 6468105 xmax: 625899 ymax: 7594111
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 94 x 5
   målemetode nøyaktighet oppdateringsdato navn                   geometry
 *      <int>       <int> <date>           <chr>               <POINT [m]>
 1         64      100000 2004-11-30       Eina         (259218.7 6729702)
 2         64      100000 2004-11-30       Kongsvinger    (333860 6676083)
 3         64      100000 2004-11-30       Moss         (253731.5 6596190)
 4         64      100000 2004-11-30       Sandefjord   (226719.1 6564929)
 5         64      100000 2004-11-30       Larvik       (215052.2 6556255)
 6         64      100000 2004-11-30       Nelaug       (130766.2 6519471)
 7         64      100000 2004-11-30       Egersund    (-24012.26 6515250)
 8         64      100000 2004-11-30       Sira         (13595.08 6505662)
 9         64      100000 2004-11-30       Arendal      (136842.8 6497335)
10         64      100000 2008-11-10       Bergen         (-31470 6733939)
# ... with 84 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Samferdsel_GML.gml}
}
\usage{
Stasjon
}
\description{
Stasjon
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
