% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Stasjon}
\alias{Stasjon}
\title{Stasjon}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 94 features and 4 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -40392.89 ymin: 6468105 xmax: 625899 ymax: 7594111
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 94 × 5
   oppdateringsdato målemetode nøyaktighet navn                   posisjon
 * <date>           <chr>            <int> <chr>               <POINT [m]>
 1 NA               gen             100000 Eina         (259218.7 6729702)
 2 NA               gen             100000 Kongsvinger    (333860 6676083)
 3 NA               gen             100000 Moss         (253731.5 6596190)
 4 NA               gen             100000 Sandefjord   (226719.1 6564929)
 5 NA               gen             100000 Larvik       (215052.2 6556255)
 6 NA               gen             100000 Nelaug       (130766.2 6519471)
 7 NA               gen             100000 Egersund    (-24012.26 6515250)
 8 NA               gen             100000 Sira         (13595.08 6505662)
 9 NA               gen             100000 Arendal      (136842.8 6497335)
10 NA               gen             100000 Bergen         (-31470 6733939)
# ℹ 84 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Samferdsel_GML.gml}
}
\usage{
Stasjon
}
\description{
Stasjon
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
