% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{StedsnavnTekst}
\alias{StedsnavnTekst}
\title{StedsnavnTekst}
\format{
\preformatted{Simple feature collection with 2030 features and 18 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -77037.69 ymin: 6448026 xmax: 1113274 ymax: 7942102
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,030 x 19
   streng oppdateringsdato datafangstdato tekstReferanseP~ tekstReferanseP~ tegnavstand
 * <chr>  <date>           <date>                    <int>            <int>       <int>
 1 Solør  2017-11-10       NA                            0                1         100
 2 Hadel~ 2017-11-10       NA                            0                2          40
 3 Land   2017-11-10       NA                            0                2         100
 4 Romer~ 2017-11-10       NA                            0                0          60
 5 Utne   2017-11-10       NA                            0                1           0
 6 Måbød~ 2017-11-10       NA                            0                1           0
 7 Veig   2017-11-10       NA                            0                0           0
 8 Kinsa~ 2017-11-10       NA                            0                0           0
 9 Jonsv~ 2017-11-10       NA                            0                0           0
10 Holts~ 2017-11-10       NA                            0                0           0
# ... with 2,020 more rows, and 13 more variables: skriftkode <int>, skrifttype <chr>,
#   referansemålestokk <int>, fulltekst <chr>, navneobjekthovedgruppe <int>,
#   navneobjektgruppe <int>, navneobjekttype <int>, språk <chr>,
#   stedsnavnsnummer <int>, språkprioritering <chr>, skrivemåtenummer <int>,
#   stedsnummer <int>, geometry <LINESTRING [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Stedsnavn_GML.gml}
}
\usage{
StedsnavnTekst
}
\description{
StedsnavnTekst
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
