% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{StedsnavnTekst}
\alias{StedsnavnTekst}
\title{StedsnavnTekst}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2055 features and 17 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -113650.4 ymin: 6448026 xmax: 1113274 ymax: 7942097
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,055 × 18
   streng     oppdateringsdato tekstReferansePunktN…¹ tekstReferansePunktØst tegnavstand
 * <chr>      <date>                            <int>                  <int>       <int>
 1 Solør      2021-11-11                            0                      1         100
 2 Hadeland   2021-11-11                            0                      2          40
 3 Land       2021-11-11                            0                      2         100
 4 Romerike   2021-11-11                            0                      0          60
 5 Utne       2021-11-11                            0                      1           0
 6 Måbødalen  2021-11-11                            0                      1           0
 7 Veig       2021-11-11                            0                      0           0
 8 Kinsarvik  2021-11-11                            0                      0           0
 9 Jonsvatnet 2021-11-11                            0                      0           0
10 Fjergen    2021-11-11                            0                      0           0
# ℹ 2,045 more rows
# ℹ abbreviated name: ¹​tekstReferansePunktNord
# ℹ 13 more variables: skriftkode <int>, skrifttype <chr>, referansemålestokk <int>,
#   fulltekst <chr>, navneobjekthovedgruppe <chr>, navneobjektgruppe <chr>,
#   navneobjekttype <chr>, språk <chr>, stedsnavnnummer <int>, språkprioritering <chr>,
#   stedsnummer <int>, skrivemåtenummer <int>, geometri <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Stedsnavn_GML.gml}
}
\usage{
StedsnavnTekst
}
\description{
StedsnavnTekst
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
