% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Terrengpunkt}
\alias{Terrengpunkt}
\title{Terrengpunkt}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 150 features and 5 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -2948 ymin: 6531310 xmax: 1105697 ymax: 7901514
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 150 × 6
   målemetode nøyaktighet medium oppdateringsdato høyde          posisjon
 *      <int>       <int> <chr>  <date>           <int>       <POINT [m]>
 1         64      100000 T      NA                 474 (1007557 7901514)
 2         64      100000 T      NA                 481 (1045873 7895568)
 3         64      100000 T      NA                 725  (997830 7883080)
 4         64      100000 T      NA                 521  (920170 7870051)
 5         64      100000 T      NA                 634  (868065 7865143)
 6         64      100000 T      NA                 545 (1055448 7860040)
 7         64      100000 T      NA                 603  (942777 7824719)
 8         64      100000 T      NA                1067  (927620 7805438)
 9         64      100000 T      NA                 405 (1105697 7794891)
10         64      100000 T      NA                1045  (896515 7793149)
# ℹ 140 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Hoyde_GML.gml}
}
\usage{
Terrengpunkt
}
\description{
Terrengpunkt
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
