% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Territorialgrense}
\alias{Territorialgrense}
\title{Territorialgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 78 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 78 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64      100000 2023-12-18                  (434193.8 7604043, 438606 7611106)
 2         64      100000 2023-12-18                (171068.4 6489902, 152049.3 6471914)
 3         64      100000 2023-12-18       (-88042.44 6593395, -93288.96 6598872, -9396…
 4         64      100000 2023-12-18       (-72652.05 6558389, -73201.83 6563624, -7325…
 5         64      100000 2023-12-18       (417975 7588030, 428888.3 7595550, 434193.8 …
 6         64      100000 2023-12-18                 (63609.57 7025260, 72502.8 7030219)
 7         64      100000 2023-12-18       (72502.8 7030219, 91125.77 7040603, 92393.5 …
 8         64      100000 2023-12-18                (247755.5 7188752, 253149.8 7193866)
 9         64      100000 2023-12-18       (380921.3 7531107, 386002.7 7545389, 386358.…
10         64      100000 2023-12-18       (583180.8 7746329, 592630.2 7765234, 593097.…
# ℹ 68 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Territorialgrense
}
\description{
Territorialgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
