% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Territorialgrense}
\alias{Territorialgrense}
\title{Territorialgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 78 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 78 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000          (701859.7 7853461, 726452.5 7865517)
 2 2023-12-18       gen             100000 (764795.7 7891994, 770308.5 7896821, 770685.…
 3 2023-12-18       gen             100000 (72502.8 7030219, 91125.77 7040603, 92393.5 …
 4 2023-12-18       gen             100000          (296959.8 7271567, 298752.7 7275338)
 5 2023-12-18       gen             100000        (-21040.29 6465214, -21808.94 6466105)
 6 2023-12-18       gen             100000          (181722.6 6499979, 171068.4 6489902)
 7 2023-12-18       gen             100000 (796042.5 7918378, 808944.4 7929181, 809725.…
 8 2023-12-18       gen             100000        (-21808.94 6466105, -36166.04 6482741)
 9 2023-12-18       gen             100000 (-70005.05 6540078, -71381.97 6547126, -7155…
10 2023-12-18       gen             100000 (548801.2 7721400, 565026.7 7729633, 579406.…
# ℹ 68 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000AdministrativeOmrader_GML.gml}
}
\usage{
Territorialgrense
}
\description{
Territorialgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
