% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Tettbebyggelse}
\alias{Tettbebyggelse}
\title{Tettbebyggelse}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 886 features and 3 fields (with 824 geometries empty)
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -58767 ymin: 6455265 xmax: 820895 ymax: 7786615
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 886 × 4
   oppdateringsdato målemetode nøyaktighet                                        område
 * <date>           <chr>            <int>                                 <POLYGON [m]>
 1 2017-01-06       <NA>                NA ((168801 6965567, 168802.8 6965349, 168806 6…
 2 2017-01-06       <NA>                NA ((651783 7737587, 651453.8 7737928, 651312.4…
 3 2017-01-06       <NA>                NA ((475827 7465979, 475576 7465753, 474098 746…
 4 2017-01-06       <NA>                NA ((105981 6980952, 105981 6981402, 105981 698…
 5 2017-01-06       <NA>                NA ((331609 7151102, 331777.1 7150347, 331852 7…
 6 2017-01-06       <NA>                NA ((287507 6750012, 287367.3 6750093, 287239.6…
 7 2017-01-06       <NA>                NA ((44136 6956284, 45104 6956434, 45404 695669…
 8 2017-01-06       <NA>                NA ((255261 6780785, 255509.2 6780884, 256419.3…
 9 2017-01-06       <NA>                NA ((-42509 6733371, -42326.54 6733119, -41486 …
10 2017-01-06       <NA>                NA ((-42709 6732572, -43119.84 6732679, -43987 …
# ℹ 876 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Arealdekke_GML.gml}
}
\usage{
Tettbebyggelse
}
\description{
Tettbebyggelse
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
