% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{VegSenterlinje}
\alias{VegSenterlinje}
\title{VegSenterlinje}
\format{
\preformatted{Simple feature collection with 2427 features and 8 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -59361.73 ymin: 6455017 xmax: 1108413 ymax: 7937309
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,427 x 9
   målemetode nøyaktighet medium oppdateringsdato vegkategori vegstatus vegnummer
 *      <int>       <int> <chr>  <date>           <chr>       <chr>         <int>
 1         64      100000 T      2008-10-16       E           V                39
 2         64      100000 T      2008-10-16       E           V                 6
 3         64      100000 T      2008-10-16       E           V               134
 4         64      100000 T      2008-10-16       E           V               134
 5         64      100000 T      2008-10-16       E           V               134
 6         64      100000 T      2008-10-16       E           V               134
 7         64      100000 T      2008-10-16       E           V                 6
 8         64      100000 T      2008-10-16       E           V                 6
 9         64      100000 T      2008-10-16       E           V               134
10         64      100000 T      2008-10-16       E           V               134
# ... with 2,417 more rows, and 2 more variables: motorvegtype <chr>,
#   geometry <LINESTRING [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N2000Samferdsel_GML.gml}
}
\usage{
VegSenterlinje
}
\description{
VegSenterlinje
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
